rename test file
